Fix3

+++ tasks/seq/sdobnov_v_complex_ccs_matrix_mult/func_tests/main.cpp
@@ -35,7 +35,7 @@ TEST(sdobnov_v_complex_ccs_matrix_mult_seq, Multiply_EmptyMatrices) {
   task.RunImpl();
   task.PostProcessingImpl();
 
-  ASSERT_EQ(static_cast<size_t>(result.values.size()), 0);
+  ASSERT_EQ(static_cast<const int>(result.values.size()), 0);
   ASSERT_EQ(result.row_i.size(), 0);
   for (int p : result.col_p) {
     ASSERT_EQ(p, 0);
